CKAN 2.11 support
Mainly fixin plugins loading in CKAN>=2.11.

The startup improvements in ckan/ckan#8219 meant that the configuration
is not always updated automatically and so if new plugins are added via
CKAN__PLUGINS we need to load them explicitly (using `plugin_loaded()`
to make sure we don't enter a loop)

The tests were a mess, and failing for 2.10. Apart from adapting them to
the new behaviour, I had to drop the DB URL configs as there was no way
of disabling the datastore plugins and/or mocking it consistently across
CKAN versions